back/requirements.txt
Django>=2.0,<3.0
psycopg2-binary>=2.8
pylint==2.5.3
djangorestframework
markdown>=3.0.0
django-filter
django-celery-results==1.2.1
django-celery-beat==2.0.0
django-crispy-forms==1.9.2
numpy
pandas
celery
redis